Adding arguments, small changes

+++ src/main.py
@@ -1,9 +1,10 @@
 import threading
 import time
-from src.Utils.DataHandler import Dataset
+from Utils.DataHandler import Dataset
 from recognise_cv2 import VideoFR
-from src.Utils.server import Server
+from Utils.server import Server
 import cv2
+import sys
 
 
 class Main:
@@ -15,6 +16,8 @@ class Main:
 
     CLEAR_UNKNOWNS_ON_LOAD = True
 
+    WITH_SERVER = True
+
     @staticmethod
     def start():
         """ Start the video capture and the server in two different threads
@@ -22,14 +25,11 @@ class Main:
 
         # Start the video capture
         video = threading.Thread(target=Main.run)
-        # Start the server to receive the requests
-
-        # To avoid circular import
-
-        server = threading.Thread(target=Server().start)
-
         video.start()
-        server.start()
+        if Main.WITH_SERVER:
+            # Start the server to receive the requests
+            server = threading.Thread(target=Server().start)
+            server.start()
 
     @staticmethod
     def run():
@@ -52,23 +52,31 @@ class Main:
 
                 # Here we use the list just assigned to pass by reference the image and not have to return it
                 if videofr_instance.capture_faces(images):
-                    cv2.imshow("Face Recognition", images[0])
+                    if VideoFR.DEBUG_MODE:
+                        cv2.imshow("Face Recognition", images[0])
 
                 # Reset frame number
                 if videofr_instance.frame_number == 100:
                     videofr_instance.frame_number = 0
-                    cv2.destroyAllWindows()
+                    if VideoFR.DEBUG_MODE:
+                        cv2.destroyAllWindows()
 
-                # Press q to quit
-                if cv2.waitKey(1) == ord('q') & 0xFF:
-                    break
-            # Sleep for 5 seconds to not overload the CPU
-            time.sleep(5)
+            # Press q to quit
+            if cv2.waitKey(1) == ord('q') & 0xFF:
+                break
+        # Sleep for 5 seconds to not overload the CPU
+        time.sleep(5)
 
-        cv2.destroyAllWindows()
+    cv2.destroyAllWindows()
 
-    # The Video Capture for FR check every 10 frames (1 second) only if there's a face detected
+# The Video Capture for FR check every 10 frames (1 second) only if there's a face detected
 
 
 if __name__ == "__main__":
+    if "--debug" in sys.argv:
+        VideoFR.DEBUG_MODE = True
+    if "--clear" in sys.argv:
+        Main.CLEAR_UNKNOWNS_ON_LOAD = True
+    if "--no-server" in sys.argv:
+        Main.WITH_SERVER = False
     Main.start()
